# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/dorm-owner/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "dorm-owner"
                ],
                "summary": "get dorm owner profile",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Dorm Owner ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.DormOwner"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "dorm-owner"
                ],
                "summary": "update dorm owner",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Dorm Owner ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "dorm owner update",
                        "name": "data",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateDormOwnerDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.DormOwner"
                        }
                    }
                }
            }
        },
        "/dorm-owner/{id}/bank-account": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "dorm-owner"
                ],
                "summary": "update bank account",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Dorm Owner ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "bank account update",
                        "name": "data",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateBankAccountDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.DormOwner"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "dorm-owner"
                ],
                "summary": "delete bank account",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Dorm Owner ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.DormOwner"
                        }
                    }
                }
            }
        },
        "/dorm-owner/{id}/picture": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "dorm-owner"
                ],
                "summary": "upload picture",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Dorm Owner ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "profile picture",
                        "name": "profile_picture",
                        "in": "formData"
                    },
                    {
                        "type": "file",
                        "description": "bank qr",
                        "name": "bank_qr",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.DormOwner"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.ErrorResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
        "dto.UpdateBankAccountDTO": {
            "type": "object",
            "required": [
                "account_name",
                "account_number",
                "bank"
            ],
            "properties": {
                "account_name": {
                    "type": "string"
                },
                "account_number": {
                    "type": "string"
                },
                "bank": {
                    "type": "string"
                }
            }
        },
...
        "dto.UpdateDormOwnerDTO": {
            "type": "object",
            "required": [
                "firstname",
                "lastname",
                "line_id",
                "phone_number"
            ],
            "properties": {
                "firstname": {
                    "type": "string"
                },
                "lastname": {
                    "type": "string"
                },
                "line_id": {
                    "type": "string"
                },
                "phone_number": {
                    "type": "string"
                }
            }
        },
...
        "model.AllDormFacility": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                }
            }
        },
        "model.AllRoomFacility": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                }
            }
        },
        "model.BankAccount": {
            "type": "object",
            "properties": {
                "account_name": {
                    "type": "string"
                },
                "account_number": {
                    "type": "string"
                },
                "bank": {
                    "type": "string"
                },
                "bank_qr_url": {
                    "type": "string"
                }
            }
        },
        "model.Dorm": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "facilities": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.AllDormFacility"
                    }
                },
                "id": {
                    "type": "integer"
                },
                "lat": {
                    "type": "number"
                },
                "long": {
                    "type": "number"
                },
                "name": {
                    "type": "string"
                },
                "nearby_locations": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.NearbyLocation"
                    }
                },
                "pictures": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.DormPicture"
                    }
                },
                "rooms": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Room"
                    }
                },
                "rules": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "zone": {
                    "type": "string"
                }
            }
        },
        "model.DormOwner": {
            "type": "object",
            "properties": {
                "bank_account": {
                    "$ref": "#/definitions/model.BankAccount"
                },
                "dorms": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Dorm"
                    }
                },
                "email": {
                    "type": "string"
                },
                "firstname": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "lastname": {
                    "type": "string"
                },
                "line_id": {
                    "type": "string"
                },
                "phone_number": {
                    "type": "string"
                },
                "picture_url": {
                    "type": "string"
                }
            }
        },
        "model.DormPicture": {
            "type": "object",
            "properties": {
                "picture_url": {
                    "type": "string"
                }
            }
        },
...
        "model.Location": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "lat": {
                    "type": "number"
                },
                "long": {
                    "type": "number"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "model.NearbyLocation": {
            "type": "object",
            "properties": {
                "distance": {
                    "type": "number"
                },
                "location": {
                    "$ref": "#/definitions/model.Location"
                }
            }
        },
...
        "model.Room": {
            "type": "object",
            "properties": {
                "available_from": {
                    "type": "string"
                },
                "capacity": {
                    "type": "integer"
                },
                "description": {
                    "type": "string"
                },
                "dorm_id": {
                    "type": "integer"
                },
                "facilities": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.AllRoomFacility"
                    }
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "pictures": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.RoomPicture"
                    }
                },
                "price": {
                    "type": "integer"
                },
                "size": {
                    "type": "number"
                }
            }
        },
...
        "model.RoomPicture": {
            "type": "object",
            "properties": {
                "picture_url": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}